char database interfaces and txt implementation, test vehicle for the parameter class
git-svn-id: http://eathena.googlecode.com/svn/branches/eapp@3434 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/common/timer.h
@@ -7,7 +7,7 @@
 
 
 #define BASE_TICK 5
-#define DIFF_TICK(a,b) (((ssize_t)(a)-(ssize_t)(b)))
+#define DIFF_TICK(a,b) (((long)(a)-(long)(b)))
 
 
 extern time_t start_time;
